backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0",
    "dotenv>=0.9.9",
    "openai>=2.13.0",
    "pyjwt>=2.10.1",
]
